requirements.txt
Pillow>=12.0.0,<13.0.0
pydantic>=2.0.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
nonebot2>=2.3.0,<3.0.0
nonebot-adapter-onebot>=2.4.0,<3.0.0
httpx>=0.23.0,<1.0.0
